backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
opencv-python==4.9.0.80
mediapipe==0.10.32
deepface==0.0.87
pyjwt==2.8.0
websockets==12.0
hypothesis==6.98.0
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-mock==3.15.1
httpx<0.28
python-multipart==0.0.6
cryptography==42.0.2
python-dotenv==1.0.0
tensorflow==2.20.0
tf-keras==2.20.1
